pyproject.toml
[tool.poetry]
name = "44-advanced-contract"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "44"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
langchain = "^0.3.4"
langchainhub = "^0.1.21"
black = "^24.10.0"
isort = "^5.13.2"
langchain-openai = "^0.2.3"
beautifulsoup4 = "^4.12.3"
langgraph = "^0.2.39"
pytest = "^8.3.3"
qdrant-client = "^1.12.0"
langchain-qdrant = "^0.1.4"
langchain-community = "^0.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
